fix XRWebGLLayer typing, move renderview to separate file
