+++ source/webgl-operate.slim.ts
@@ -66,4 +66,5 @@ export import gl_matrix_extensions = root_gl_matrix_extensions;
 import * as root_tuples from './tuples';
 export import tuples = root_tuples;
 
-export { supportsXR, RenderView, XRController } from './xrcontroller';
+export { supportsXR, XRController } from './xrcontroller';
+export { RenderView } from './renderview';
